Rename file and service for consistency
